+++ src/aks-preview/azext_aks_preview/vendored_sdks/azure_mgmt_preview_aks/aio/operations/_managed_cluster_snapshots_operations.py
@@ -39,9 +39,9 @@ from ...operations._managed_cluster_snapshots_operations import (
 )
 from .._configuration import ContainerServiceClientConfiguration
 
-List = list
 T = TypeVar("T")
 ClsType = Optional[Callable[[PipelineResponse[HttpRequest, AsyncHttpResponse], T, dict[str, Any]], Any]]
+List = list
 
 
 class ManagedClusterSnapshotsOperations: