+++ src/aks-preview/azext_aks_preview/vendored_sdks/azure_mgmt_preview_aks/aio/operations/_managed_clusters_operations.py
@@ -69,9 +69,9 @@ from ...operations._managed_clusters_operations import (
 )
 from .._configuration import ContainerServiceClientConfiguration
 
-List = list
 T = TypeVar("T")
 ClsType = Optional[Callable[[PipelineResponse[HttpRequest, AsyncHttpResponse], T, dict[str, Any]], Any]]
+List = list
 
 
 class ManagedClustersOperations:  # pylint: disable=too-many-public-methods