+++ src/aks-preview/azext_aks_preview/vendored_sdks/azure_mgmt_preview_aks/aio/operations/_private_endpoint_connections_operations.py
@@ -38,9 +38,9 @@ from ...operations._private_endpoint_connections_operations import (
 )
 from .._configuration import ContainerServiceClientConfiguration
 
-List = list
 T = TypeVar("T")
 ClsType = Optional[Callable[[PipelineResponse[HttpRequest, AsyncHttpResponse], T, dict[str, Any]], Any]]
+List = list
 
 
 class PrivateEndpointConnectionsOperations: